Rename masonry library file
